Extract coauthors package from main
Move co-author tracking logic (AppendCoauthorsToSquashMsg,
CollectCoauthorsFromWipCommits, CreateCommitMessage and helper
functions) into a dedicated coauthors/ package. The gitUserEmail()
dependency is now injected as a parameter instead of called internally.

Unit tests moved to coauthors/coauthors_test.go, integration test
TestStartDoneCoAuthors remains in root package.
