+++ coauthors_test.go
@@ -44,29 +44,3 @@ func TestStartDoneCoAuthors(t *testing.T) {
 	// include everyone else in commit order after removing duplicates
 	assertOutputContains(t, &output, "\nCo-authored-by: bob <bob@example.com>\nCo-authored-by: alice <alice@example.com>\nCo-authored-by: localother <localother@example.com>\n")
 }
-
-func TestCreateCommitMessage(t *testing.T) {
-	equals(t, `
-
-# automatically added all co-authors from WIP commits
-# add missing co-authors manually
-Co-authored-by: Alice <alice@example.com>
-Co-authored-by: Bob <bob@example.com>
-`, createCommitMessage([]Author{"Alice <alice@example.com>", "Bob <bob@example.com>"}))
-}
-
-func TestSortByLength(t *testing.T) {
-	slice := []string{"aa", "b"}
-
-	sortByLength(slice)
-
-	equals(t, []string{"b", "aa"}, slice)
-}
-
-func TestRemoveDuplicateValues(t *testing.T) {
-	slice := []string{"aa", "b", "c", "b"}
-
-	actual := removeDuplicateValues(slice)
-
-	equals(t, []string{"aa", "b", "c"}, actual)
-}
